Implemented Upgrade Scene (Skeleton)
Access Upgrade Scene by clicking on fridge in Home Base. Can't test that for now due to a bug, I can't move in the scene

From Upgrade scene, there is an exit button to return to home base

I implemented the basic structure of the Upgrade Scene without the logic behind it.